requirements.txt
ultralytics
sahi
PyMuPDF
opencv-python
pandas
tqdm
pytest
black
pyyaml

